requirements.txt
openai
dotenv
tqdm
inquirer
gTTS
python-vlc
playsound